Addresses #229 - Explicit cast needed with VS when using += to add a char* to an int.
git-svn-id: http://developer.mangband.org/svn/mangband/trunk@733 802d84f3-b497-254e-bede-06c0a5a3456a

+++ src/server/netserver.c
@@ -1426,7 +1426,7 @@ int process_pending_commands(int ind)
 			type = (connp->r.ptr[0] & 0xFF);
 			result = (*receive_tbl[type])(ind);
 			if ((connp->r.ptr - last_pos) > 0) 
-				data_advance += (connp->r.ptr - last_pos);
+				data_advance += (int)(connp->r.ptr - last_pos);
 			connp->start = turn;
 			if (result == 0)
 				return TRUE;
